// src/types/IntroDataType.tsx
export interface FestivalIntroInfo {
  contentid: string;
  contenttypeid: string;
  sponsor1: string;
  sponsor1tel: string;
  sponsor2: string;
  sponsor2tel: string;
  eventenddate: string;
  playtime: string;
  eventplace: string;
  eventhomepage: string;
  agelimit: string;
  bookingplace: string;
  placeinfo: string;
  subevent: string;
  program: string;
  eventstartdate: string;
  usetimefestival: string;
  discountinfofestival: string;
  spendtimefestival: string;
  festivalgrade: string;
}

export interface LodgmentIntroInfo {
  contentid: string;
  contenttypeid: string;
  goodstay: string;
  benikia: string;
  hanok: string;
  roomcount: string;
  roomtype: string;
  refundregulation: string;
  checkintime: string;
  checkouttime: string;
  chkcooking: string;
  seminar: string;
  sports: string;
  sauna: string;
  beauty: string;
  beverage: string;
  karaoke: string;
  barbecue: string;
  campfire: string;
  bicycle: string;
  fitness: string;
  publicpc: string;
  publicbath: string;
  subfacility: string;
  foodplace: string;
  reservationurl: string;
  pickup: string;
  infocenterlodging: string;
  parkinglodging: string;
  reservationlodging: string;
  scalelodging: string;
  accomcountlodging: string;
}

export interface AttractionIntroInfo {
  contentid: string;
  contenttypeid: string;
  heritage1: string;
  heritage2: string;
  heritage3: string;
  infocenter: string;
  opendate: string;
  restdate: string;
  expguide: string;
  expagerange: string;
  accomcount: string;
  useseason: string;
  usetime: string;
  parking: string;
  chkbabycarriage: string;
  chkpet: string;
  chkcreditcard: string;
}